fixed Crawler.java fixed Indexer.java added IndexerIO.java removed old IndexedDatabase.java
considering adding new indexing storage cuz it currently takes up to 3 gigs of ram when used with 6000k websites

+++ src/main/java/uni/apt/engine/WordRecord.java
@@ -6,7 +6,7 @@ public class WordRecord implements Serializable {
     public int pos;
     public int type;
     public int tagIndex;
-    public int paragraphIndex;
+    public long paragraphIndex;
 
     public static final int PARAGRAPH = 0;
     public static final int HEADER    = 1;
